@@ -1,7 +1,8 @@
 package com.ryan.rv_gallery;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  *
